+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # CHANGELOG
 
 
+## v0.4.15
+* Update to Bootstrap 4.1
+* Update SweetAlert Spinner to include Title
+* Add ability to pass className and attributes to ScrollableContainer component
+* Add tablet and desktop icons for preview buttons
+* Update Sweet Alert logout buttons in admin-ui
+* Set default size to small for checkbox, radio and switch
+* Add sd size to checkbox, radio and switch
+
+
 ## v0.4.14
 * Add SweetAlert local component on top of SweetAlert2
 * Match SweetAlert and Modal styling so you can use interchangably